+++ inst/doc/vignette_for_IRTest.R
@@ -1,4 +1,4 @@
-## ----include = FALSE----------------------------------------------------------
+## ---- include = FALSE---------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
@@ -22,13 +22,13 @@ data <- Alldata$data_D
 theta <- Alldata$theta
 colnames(data) <- paste0("item", 1:15)
 
-## ----results='hide', message=FALSE--------------------------------------------
+## ---- results='hide', message=FALSE-------------------------------------------
 Mod1 <- IRTest_Dich(data = data,
                     model = 2,
                     latent_dist = "LLS",
                     h=4)
 
-## ----message=FALSE, fig.align='center', fig.height=6, fig.width=6-------------
+## ---- message=FALSE, fig.align='center', fig.height=6, fig.width=6------------
 ### Summary
 summary(Mod1)
 
@@ -49,7 +49,7 @@ abline(b=1, a=0)
 ### Standard errors of ability parameter estimates
 plot(fscore$theta, fscore$theta_se)
 
-## ----fig.align='center', fig.asp=.6, fig.width=6------------------------------
+## ---- fig.align='center', fig.asp=.6, fig.width=6-----------------------------
 plot(Mod1, mapping = aes(colour="Estimated"), linewidth = 1) +
   lims(y = c(0, .75))+
   geom_line(
@@ -61,7 +61,7 @@ plot(Mod1, mapping = aes(colour="Estimated"), linewidth = 1) +
   labs(title="The estimated latent density using '2NM'", colour= "Type")+
   theme_bw()
 
-## ----fig.align='center', fig.asp=0.8, fig.width=6-----------------------------
+## ---- fig.align='center', fig.asp=0.8, fig.width=6----------------------------
 p1 <- plot_item(Mod1,1)
 p2 <- plot_item(Mod1,4)
 p3 <- plot_item(Mod1,8)
@@ -74,7 +74,7 @@ item_fit(Mod1)
 ## -----------------------------------------------------------------------------
 reliability(Mod1)
 
-## ----fig.align='center', fig.asp=.7, fig.width=6------------------------------
+## ---- fig.align='center', fig.asp=.7, fig.width=6-----------------------------
 set.seed(1)
 selected_examinees <- sample(1:1000,6)
 post_sample <- 
@@ -92,7 +92,7 @@ ggplot(data=post_sample, mapping=aes(x=X))+
   facet_wrap(~ID, ncol=2)+
   theme_bw()
 
-## ----fig.align='center', fig.asp=.8, fig.width=6------------------------------
+## ---- fig.align='center', fig.asp=.8, fig.width=6-----------------------------
 ggplot()+
   stat_function(
     fun = inform_f_test,
@@ -141,12 +141,12 @@ data <- Alldata$data_P
 theta <- Alldata$theta
 colnames(data) <- paste0("item", 1:14)
 
-## ----results='hide', message=FALSE--------------------------------------------
+## ---- results='hide', message=FALSE-------------------------------------------
 Mod1 <- IRTest_Poly(data = data,
                     model = "GRM",
                     latent_dist = "KDE")
 
-## ----message=FALSE, fig.align='center', fig.height=6, fig.width=6-------------
+## ---- message=FALSE, fig.align='center', fig.height=6, fig.width=6------------
 ### Summary
 summary(Mod1)
 
@@ -167,7 +167,7 @@ abline(b=1, a=0)
 ### Standard errors of ability parameter estimates
 plot(fscore$theta, fscore$theta_se)
 
-## ----fig.align='center', fig.asp=.6, fig.width=6------------------------------
+## ---- fig.align='center', fig.asp=.6, fig.width=6-----------------------------
 plot(Mod1, mapping = aes(colour="Estimated"), linewidth = 1) +
   stat_function(
     fun = dist2,
@@ -178,7 +178,7 @@ plot(Mod1, mapping = aes(colour="Estimated"), linewidth = 1) +
   labs(title="The estimated latent density using '2NM'", colour= "Type")+
   theme_bw()
 
-## ----fig.align='center', fig.asp=0.8, fig.width=6-----------------------------
+## ---- fig.align='center', fig.asp=0.8, fig.width=6----------------------------
 p1 <- plot_item(Mod1,1)
 p2 <- plot_item(Mod1,4)
 p3 <- plot_item(Mod1,8)
@@ -191,7 +191,7 @@ item_fit(Mod1)
 ## -----------------------------------------------------------------------------
 reliability(Mod1)
 
-## ----fig.align='center', fig.asp=.7, fig.width=6------------------------------
+## ---- fig.align='center', fig.asp=.7, fig.width=6-----------------------------
 set.seed(1)
 selected_examinees <- sample(1:1000,6)
 post_sample <- 
@@ -209,7 +209,7 @@ ggplot(data=post_sample, mapping=aes(x=X))+
   facet_wrap(~ID, ncol=2)+
   theme_bw()
 
-## ----fig.align='center', fig.asp=.8, fig.width=6------------------------------
+## ---- fig.align='center', fig.asp=.8, fig.width=6-----------------------------
 ggplot()+
   stat_function(
     fun = inform_f_test,
@@ -250,8 +250,8 @@ ggplot()+
   theme_bw()
 
 ## -----------------------------------------------------------------------------
-Alldata <- DataGeneration(N=500,
-                          nitem_C = 4,
+Alldata <- DataGeneration(N=1000,
+                          nitem_C = 8,
                           latent_dist = "2NM",
                           a_l = .3, 
                           a_u = .7,
@@ -261,16 +261,13 @@ Alldata <- DataGeneration(N=500,
 
 data <- Alldata$data_C
 theta <- Alldata$theta
-colnames(data) <- paste0("item", 1:4)
+colnames(data) <- paste0("item", 1:8)
 
-## ----results='hide', message=FALSE--------------------------------------------
+## ---- results='hide', message=FALSE-------------------------------------------
 Mod1 <- IRTest_Cont(data = data,
-                    latent_dist = "2NM",
-                    range = c(-4,4),
-                    q = 81,
-                    threshold = 0.01)
+                    latent_dist = "KDE")
 
-## ----message=FALSE, fig.align='center', fig.height=6, fig.width=6-------------
+## ---- message=FALSE, fig.align='center', fig.height=6, fig.width=6------------
 ### Summary
 summary(Mod1)
 
@@ -281,14 +278,14 @@ logLik(Mod1)
 coef(Mod1)
 
 ### The estimated ability parameters
-fscore <- factor_score(Mod1, ability_method = "MLE")
+fscore <- factor_score(Mod1, ability_method = "WLE")
 plot(theta, fscore$theta)
 abline(b=1, a=0)
 
 ### Standard errors of ability parameter estimates
 plot(fscore$theta, fscore$theta_se)
 
-## ----fig.align='center', fig.asp=.6, fig.width=6------------------------------
+## ---- fig.align='center', fig.asp=.6, fig.width=6-----------------------------
 plot(Mod1, mapping = aes(colour="Estimated"), linewidth = 1) +
   lims(y = c(0, .75))+
   geom_line(
@@ -300,7 +297,7 @@ plot(Mod1, mapping = aes(colour="Estimated"), linewidth = 1) +
   labs(title="The estimated latent density using '2NM'", colour= "Type")+
   theme_bw()
 
-## ----fig.align='center', fig.asp=0.8, fig.width=6-----------------------------
+## ---- fig.align='center', fig.asp=0.8, fig.width=6----------------------------
 p1 <- plot_item(Mod1,1)
 p2 <- plot_item(Mod1,2)
 p3 <- plot_item(Mod1,3)
@@ -310,15 +307,15 @@ grid.arrange(p1, p2, p3, p4, ncol=2, nrow=2)
 ## -----------------------------------------------------------------------------
 reliability(Mod1)
 
-## ----fig.align='center', fig.asp=.7, fig.width=6------------------------------
+## ---- fig.align='center', fig.asp=.7, fig.width=6-----------------------------
 set.seed(1)
-selected_examinees <- sample(1:500,6)
+selected_examinees <- sample(1:1000,6)
 post_sample <- 
   data.frame(
-    X = rep(seq(-4,4, length.out=81),6), 
+    X = rep(seq(-6,6, length.out=121),6), 
     prior = rep(Mod1$Ak/(Mod1$quad[2]-Mod1$quad[1]), 6),
     posterior = 10*c(t(Mod1$Pk[selected_examinees,])), 
-    ID = rep(paste("examinee", selected_examinees), each=81)
+    ID = rep(paste("examinee", selected_examinees), each=121)
     )
 
 ggplot(data=post_sample, mapping=aes(x=X))+
@@ -328,7 +325,7 @@ ggplot(data=post_sample, mapping=aes(x=X))+
   facet_wrap(~ID, ncol=2)+
   theme_bw()
 
-## ----fig.align='center', fig.asp=.8, fig.width=6------------------------------
+## ---- fig.align='center', fig.asp=.8, fig.width=6-----------------------------
 ggplot()+
   stat_function(
     fun = inform_f_test,
@@ -354,6 +351,26 @@ ggplot()+
     args = list(Mod1, 4),
     mapping = aes(color="Item 4")
   )+
+  stat_function(
+    fun=inform_f_item,
+    args = list(Mod1, 5),
+    mapping = aes(color="Item 5")
+  )+
+  stat_function(
+    fun=inform_f_item,
+    args = list(Mod1, 6),
+    mapping = aes(color="Item 6")
+  )+
+  stat_function(
+    fun=inform_f_item,
+    args = list(Mod1, 7),
+    mapping = aes(color="Item 7")
+  )+
+  stat_function(
+    fun=inform_f_item,
+    args = list(Mod1, 8),
+    mapping = aes(color="Item 8")
+  )+
   lims(x=c(-6,6))+
   labs(title="Test information function", x=expression(theta), y='information')+
   theme_bw()
@@ -375,14 +392,14 @@ theta <- Alldata$theta
 colnames(DataD) <- paste0("item", 1:10)
 colnames(DataP) <- paste0("item", 1:5)
 
-## ----results='hide', message=FALSE--------------------------------------------
+## ---- results='hide', message=FALSE-------------------------------------------
 Mod1 <- IRTest_Mix(data_D = DataD,
                    data_P = DataP,
                    model_D = "2PL",
                    model_P = "GRM",
                    latent_dist = "KDE")
 
-## ----message=FALSE, fig.align='center', fig.height=6, fig.width=6-------------
+## ---- message=FALSE, fig.align='center', fig.height=6, fig.width=6------------
 ### Summary
 summary(Mod1)
 
@@ -403,7 +420,7 @@ abline(b=1, a=0)
 ### Standard errors of ability parameter estimates
 plot(fscore$theta, fscore$theta_se)
 
-## ----fig.align='center', fig.asp=.6, fig.width=6------------------------------
+## ---- fig.align='center', fig.asp=.6, fig.width=6-----------------------------
 plot(Mod1, mapping = aes(colour="Estimated"), linewidth = 1) +
   stat_function(
     fun = dist2,
@@ -414,14 +431,14 @@ plot(Mod1, mapping = aes(colour="Estimated"), linewidth = 1) +
   labs(title="The estimated latent density using '2NM'", colour= "Type")+
   theme_bw()
 
-## ----fig.align='center', fig.asp=0.8, fig.width=6-----------------------------
+## ---- fig.align='center', fig.asp=0.8, fig.width=6----------------------------
 p1 <- plot_item(Mod1,1, type="d")
 p2 <- plot_item(Mod1,4, type="d")
 p3 <- plot_item(Mod1,8, type="d")
 p4 <- plot_item(Mod1,10, type="d")
 grid.arrange(p1, p2, p3, p4, ncol=2, nrow=2)
 
-## ----fig.align='center', fig.asp=0.8, fig.width=6-----------------------------
+## ---- fig.align='center', fig.asp=0.8, fig.width=6----------------------------
 p1 <- plot_item(Mod1,1, type="p")
 p2 <- plot_item(Mod1,2, type="p")
 p3 <- plot_item(Mod1,3, type="p")
@@ -434,7 +451,7 @@ item_fit(Mod1)
 ## -----------------------------------------------------------------------------
 reliability(Mod1)
 
-## ----fig.align='center', fig.asp=.7, fig.width=6------------------------------
+## ---- fig.align='center', fig.asp=.7, fig.width=6-----------------------------
 set.seed(1)
 selected_examinees <- sample(1:1000,6)
 post_sample <- 
@@ -452,7 +469,7 @@ ggplot(data=post_sample, mapping=aes(x=X))+
   facet_wrap(~ID, ncol=2)+
   theme_bw()
 
-## ----fig.align='center', fig.asp=.8, fig.width=8------------------------------
+## ---- fig.align='center', fig.asp=.8, fig.width=8-----------------------------
 ggplot()+
   stat_function(
     fun = inform_f_test,
@@ -500,7 +517,7 @@ data <- DataGeneration(N=1000,
                        sd_ratio = 2,
                        prob = 0.3)$data_D
 
-## ----results='hide', message=FALSE--------------------------------------------
+## ---- results='hide', message=FALSE-------------------------------------------
 model_fits <- list()
 model_fits[[1]] <- IRTest_Dich(data)
 model_fits[[2]] <- IRTest_Dich(data, latent_dist = "EHM")